+++ src/app/interfaces/Advisor.ts
@@ -3,6 +3,7 @@ import { MarketplaceItem } from "celeste-api-types"
 import { Vendor } from "./Vendor"
 
 export interface AdvisorRarity {
+  id: string
   description: string
   icon: number,
 }
